refactor: update ModularSynth to use computeBlock() and improve BlockContext handling

+++ PatchCore/src/main/cpp/include/patchcore/module/buffer/BlockContext.hpp
@@ -1,10 +1,12 @@
 #ifndef PATCHCORE_BLOCKCONTEXT_HPP
 #define PATCHCORE_BLOCKCONTEXT_HPP
 
+#include <cstdint>
+
 struct BlockContext {
     int blockSize = 0;
     int sampleRate = 0;
-    long long blockStartSample = 0;
+    int64_t blockStartSample = 0;
     double blockStartTimeUs = 0.0;
 };
